+++ tests/dummy/app/router.js
@@ -7,6 +7,8 @@ var Router = Ember.Router.extend({
 
 Router.map(function() {
     this.route('dropdown');
+    this.route('button');
+    this.route('forms');
     this.route('accordion');
 });