+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+<a name"0.1.2"></a>
+### 0.1.2 (2015-04-22)
+
+#### Features
+
+* **I18n:** support for ember-i18n added: selected components support translateable properties
+
+<a name"0.1.1"></a>
+### 0.1.1 (2015-04-21)
+
+#### Bug Fixes
+
+* **FormGroup:** fixed wrong form group validation class in Ember 1.11 ([ff72909e](https://github.com/kaliber5/ember-bootstrap/commit/ff72909e))
+
+
+<a name"0.1.0"></a>
+## 0.1.0 (2015-04-17)
+
+#### Features
+
+New form classes to support creating powerful forms:
+
+* **Form:** base form class
+* **FormGroup:** component for Bootstrap form-group elements with validation rendering
+* **FormElement:** extends FormGroup to implement automatic markup generation and model based validations
+
+<a name"0.0.9"></a>
+### 0.0.9 (2015-04-08)
+
+#### Bug Fixes
+
+* **ButtonGroup:**
+  * initial value did not set active buttons properly
+
+<a name"0.0.8"></a>
+### 0.0.8 (2015-04-08)
+
+
+#### Bug Fixes
+
+* **ButtonGroup:**
+  * fixed radio button being clicked twice to deactivate itself ([6175e12d](https://github.com/kaliber5/ember-bootstrap/commit/6175e12d))
+  * fixed broken button group value added separate button page to dummy app ([2c2dafc0](https://github.com/kaliber5/ember-bootstrap/commit/2c2dafc0))
+* **bower:** fixed missing bootstrap dependency in bower.json. Did not affect the host ember  ([5c00c2b9](https://github.com/kaliber5/ember-bootstrap/commit/5c00c2b9))
+
 <a name"0.0.7"></a>
 ### 0.0.7 (2015-03-14)